endpoints/midpoint-federated-oauth-callbacks/requirements.txt
Authlib==1.6.9
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
idna==3.11
pycparser==3.0
PyJWT==2.12.1
requests==2.33.1
urllib3==2.6.3
